feat: add PostsWithEmbeds component
There are a couple of lexicons on the AT protocol, and embeds is one of them. A post can have multiple embeds in it. The common one is an image embed where people make a post on Bluesky with a caption and an attached image. We have quoted posts too, and many more.

This commits introduces three of those embed $types, a record, external, and image embeds. As time goes one, we'll include more of these.